+++ tests/testthat/_problems/test_ivregranks_vcov-8.R
@@ -1,0 +1,10 @@
+# Extracted from test_ivregranks_vcov.R:8
+
+# setup ------------------------------------------------------------------------
+library(testthat)
+test_env <- simulate_test_env(package = "csranks", path = "..")
+attach(test_env, warn.conflicts = FALSE)
+
+# test -------------------------------------------------------------------------
+model <- ivregranks(r(mpg) ~ r(hp) + cyl | r(disp) + cyl, data = mtcars)
+sumr <- summary(model)